Se ajusta el tipo de datos de la edad a tipo byte

+++ src/test/java/co/edu/uniquindio/poo/AppTest.java
@@ -25,14 +25,14 @@ public class AppTest {
     public void datosCompletos() {
         LOG.info("Inicio datosCompletos");
 
-        Estudiante estudiante = new Estudiante("Camila", "Alzate Rios", "109453264", "camila@uniquindio.edu.co", "315635674", 18);
+        Estudiante estudiante = new Estudiante("Camila", "Alzate Rios", "109453264", "camila@uniquindio.edu.co", "315635674", (byte)18);
 
         assertEquals("Camila", estudiante.nombres());
         assertEquals("Alzate Rios", estudiante.apellidos());
         assertEquals("109453264", estudiante.numeroIdentificacion());
         assertEquals("camila@uniquindio.edu.co", estudiante.correo());
         assertEquals("315635674", estudiante.telefono());
-        assertEquals(18, estudiante.edad());
+        assertEquals((byte)18, estudiante.edad());
         LOG.info("Finalización datosCompletos");
     }
 
@@ -42,7 +42,7 @@ public class AppTest {
     @Test
     public void datosNulos() {
         LOG.info("Inicio datosNulos");
-        assertThrows(Throwable.class, () -> new Estudiante(null, null, null, null, null, 18));
+        assertThrows(Throwable.class, () -> new Estudiante(null, null, null, null, null, (byte)18));
         LOG.info("Finalización datosNulos");
     }
 
@@ -52,7 +52,7 @@ public class AppTest {
     @Test
     public void datosVacios() {
         LOG.info("Inicio datosVacios");
-        assertThrows(Throwable.class, () -> new Estudiante("", "", "", "", "", 0));
+        assertThrows(Throwable.class, () -> new Estudiante("", "", "", "", "", (byte)0));
         LOG.info("Finalización datosVacios");
     }
 
@@ -62,7 +62,7 @@ public class AppTest {
     @Test
     public void edadNegativa() {
         LOG.info("Inicio datosNegativos");
-        assertThrows(Throwable.class, () -> new Estudiante("Camila", "Alzate Rios", "109453264", "camila@uniquindio.edu.co", "315635674", -18));
+        assertThrows(Throwable.class, () -> new Estudiante("Camila", "Alzate Rios", "109453264", "camila@uniquindio.edu.co", "315635674", (byte)-18));
         LOG.info("Finalización datosNegativos");
     }
 
@@ -72,7 +72,7 @@ public class AppTest {
     @Test
     public void correoInvalido() {
         LOG.info("Inicio correoInvalido");
-        assertThrows(Throwable.class, () -> new Estudiante("Camila", "Alzate Rios", "109453264", "cami", "315635674", 18));
+        assertThrows(Throwable.class, () -> new Estudiante("Camila", "Alzate Rios", "109453264", "cami", "315635674", (byte)18));
         LOG.info("Finalización correoInvalido");
     }
